@@ -11,10 +11,11 @@
     "build": "node build/build.js"
   },
   "dependencies": {
+    "bootstrap-vue": "^2.0.0-rc.27",
     "firebase": "^6.3.3",
     "inspector": "^0.5.0",
     "moment": "^2.24.0",
-    "vue": "^2.5.2",
+    "vue": "^2.6.10",
     "vue-router": "^3.0.1",
     "vue-uuid": "^1.1.1",
     "vuex": "^3.1.1"
@@ -30,6 +31,7 @@
     "babel-plugin-transform-vue-jsx": "^3.5.0",
     "babel-preset-env": "^1.3.2",
     "babel-preset-stage-2": "^6.22.0",
+    "bootstrap": "^4.3.1",
     "chalk": "^2.0.1",
     "copy-webpack-plugin": "^4.0.1",
     "css-loader": "^0.28.0",
@@ -46,10 +48,12 @@
     "file-loader": "^1.1.4",
     "friendly-errors-webpack-plugin": "^1.6.1",
     "html-webpack-plugin": "^2.30.1",
+    "jquery": "^3.4.1",
     "node-notifier": "^5.1.2",
     "node-sass": "^4.12.0",
     "optimize-css-assets-webpack-plugin": "^3.2.0",
     "ora": "^1.2.0",
+    "popper.js": "^1.15.0",
     "portfinder": "^1.0.13",
     "postcss-import": "^11.0.0",
     "postcss-loader": "^2.0.8",
