@@ -1,55 +1,108 @@
 <template>
   <div id="dashboard">
-    <router-link to="/home">Home</router-link>
-    <section v-if="!showAddsection">
-      <a @click="viewAddSection">Add new item</a>
-      <input type="text" v-model="searchterm" placeholder="Search items">
-      <input type="file" @change="onImportFilePicked">
-      <ul v-if="filteredItems.length">
-        <li v-for="item in filteredItems" :key="item.u_id" class="post">
-          <img :src="item.imageurl" height="100">
-          {{ item.id }}
-          {{ item.u_id }}
-          {{ item.code }}
-          {{ item.category }}
-          {{ item.name }}
-          {{ item.description }}
-          {{ item.price }}
-          <button
-            @click="editItem(item)"
-          >Edit</button>
-          <button @click="deleteItem(item.id)">Delete</button>
-        </li>
-      </ul>
+    <b-container v-if="!showAddsection">
+      <b-row>
+        <b-col cols="6" align-v="end">
+          <b-form-input type="text" v-model="searchterm" placeholder="Search items"/>
+        </b-col>
+        <b-col cols="3">
+          <b-button variant="primary" @click="viewAddSection">Add new item</b-button>
+        </b-col>
+        <b-col cols="3">
+          <b-button center variant="secondary" @click="onPickUploadFile">Upload Items</b-button>
+          <input type="file" hidden @change="onImportFilePicked" ref="fileUpload">
+        </b-col>
+      </b-row>
+
+      <b-table
+        v-if="filteredItems.length"
+        borderless
+        outlined
+        hover
+        :items="filteredItems"
+        :fields="fields"
+      >
+        <template slot="image" slot-scope="data">
+          <img :src="data.item.imageurl" height="100">
+        </template>
+        <template slot="actions" slot-scope="data">
+          <b-button variant="outline-primary" @click="editItem(data.item)">Edit</b-button>
+          <b-button variant="danger" @click="deleteItem(data.item.id)">Delete</b-button>
+        </template>
+      </b-table>
+
       <div v-else>
         <p class="no-results">There are currently no items</p>
       </div>
-    </section>
+    </b-container>
+
     <transition name="fade">
       <div v-if="showAddsection" class="p-modal">
         <div class="p-container">
           <a @click="closePostModal" class="close">X</a>
-          <div class="post">
-            Code :
-            <input v-model="modifiedItem.code">
-            Category :
-            <input v-model="modifiedItem.category">
-            Name :
-            <input v-model="modifiedItem.name">
-            Description :
-            <input v-model="modifiedItem.description">
-            Price :
-            <input v-model="modifiedItem.price">
-            Image :
-            <input v-model="modifiedItem.imageurl">
-            <button class="primary" @click="onPickFile">Upload Image</button>
-            <input type="file" accept="image/*" @change="onFilePicked" hidden ref="fileInput">
-            <img :src="imageUrl" height="150">
-            <button
-              v-if="modifiedItem.code && modifiedItem.category && modifiedItem.name && modifiedItem.description && modifiedItem.price"
-              @click="addItem"
-            >Save</button>
-          </div>
+          <b-form class="post">
+            <b-form-group label="Code:" label-for="code">
+              <b-form-input
+                id="input-1"
+                v-model="modifiedItem.code"
+                type="text"
+                required
+                placeholder="Code"
+              ></b-form-input>
+            </b-form-group>
+
+            <b-form-group label="Category:" label-for="category">
+              <b-form-input
+                id="category"
+                v-model="modifiedItem.category"
+                type="text"
+                required
+                placeholder="Category"
+              ></b-form-input>
+            </b-form-group>
+
+            <b-form-group label="Name:" label-for="name">
+              <b-form-input
+                id="name"
+                v-model="modifiedItem.name"
+                type="text"
+                required
+                placeholder="Name"
+              ></b-form-input>
+            </b-form-group>
+
+            <b-form-group label="Description:" label-for="description">
+              <b-form-input
+                id="description"
+                v-model="modifiedItem.description"
+                type="text"
+                placeholder="Description"
+              ></b-form-input>
+            </b-form-group>
+
+            <b-form-group label="Price:" label-for="price">
+              <b-form-input
+                id="price"
+                v-model="modifiedItem.price"
+                type="number"
+                required
+                placeholder="Price"
+              ></b-form-input>
+            </b-form-group>
+
+            <b-form-group label="Image:" label-for="imageurl">
+              <b-form-input
+                id="imageurl"
+                v-model="modifiedItem.imageurl"
+                type="text"
+                placeholder="Image URL"
+              ></b-form-input>
+              <b-button center variant="secondary" @click="onPickFile">Upload Image</b-button>
+              <input type="file" accept="image/*" @change="onFilePicked" hidden ref="fileInput">
+              <b-img :src="imageUrl"></b-img>
+            </b-form-group>
+            <b-button center variant="primary" @click="addItem">Save</b-button>
+          </b-form>
         </div>
       </div>
     </transition>
@@ -57,33 +110,34 @@
 </template>
 
 <script>
-import { mapState } from "vuex";
-import { uuid } from "vue-uuid";
-const fb = require("../firebaseConfig.js");
+import { mapState } from 'vuex'
+import { uuid } from 'vue-uuid'
+const fb = require('../firebaseConfig.js')
 export default {
-  data() {
+  data () {
     return {
-      imageUrl: "",
-      imageFile: "",
-      image: "",
-      searchterm: "",
+      fields: ['image', 'code', 'name', 'price', 'description', 'actions'],
+      imageUrl: '',
+      imageFile: '',
+      image: '',
+      searchterm: '',
       modifiedItem: {
-        id: "",
-        code: "",
-        name: "",
-        description: "",
-        category: "",
-        u_id: "",
-        imageurl: "",
-        price: ""
+        id: '',
+        code: '',
+        name: '',
+        description: '',
+        category: '',
+        u_id: '',
+        imageurl: '',
+        price: ''
       },
       showAddsection: false,
       isEdit: false
-    };
+    }
   },
   computed: {
-    ...mapState(["currentUser", "items"]),
-    filteredItems: function() {
+    ...mapState(['currentUser', 'items']),
+    filteredItems: function () {
       return this.items.filter(item => {
         return (
           item.code
@@ -98,50 +152,50 @@ export default {
             .toLowerCase()
             .trim()
             .match(this.searchterm.toLowerCase().trim())
-        );
-      });
+        )
+      })
     }
   },
   methods: {
-    addItem() {
+    addItem () {
       if (this.imageUrl) {
         fb.storage
-          .ref("items/" + this.imageFile.name)
+          .ref('items/' + this.imageFile.name)
           .put(this.imageFile)
           .then(filedata => {
             filedata.ref.getDownloadURL().then(url => {
-              this.modifiedItem.imageurl = url;
-              this.imageUrl = "";
-              this.imageFile = "";
-              this.image = "";
-              this.saveItem();
-            });
-          });
+              this.modifiedItem.imageurl = url
+              this.imageUrl = ''
+              this.imageFile = ''
+              this.image = ''
+              this.saveItem()
+            })
+          })
       } else {
-        this.saveItem();
+        this.saveItem()
       }
     },
 
-    saveItem() {
+    saveItem () {
       if (this.isEdit) {
         fb.itemsCollection
           .doc(this.modifiedItem.id)
           .set(this.modifiedItem)
           .then(doc => {
-            this.modifiedItem.id = "";
-            this.modifiedItem.code = "";
-            this.modifiedItem.name = "";
-            this.modifiedItem.description = "";
-            this.modifiedItem.category = "";
-            this.modifiedItem.u_id = "";
-            this.modifiedItem.imageurl = "";
-            this.modifiedItem.price = "";
-            this.showAddsection = false;
-            this.isEdit = false;
+            this.modifiedItem.id = ''
+            this.modifiedItem.code = ''
+            this.modifiedItem.name = ''
+            this.modifiedItem.description = ''
+            this.modifiedItem.category = ''
+            this.modifiedItem.u_id = ''
+            this.modifiedItem.imageurl = ''
+            this.modifiedItem.price = ''
+            this.showAddsection = false
+            this.isEdit = false
           })
           .catch(err => {
-            console.log(err);
-          });
+            console.log(err)
+          })
       } else {
         fb.itemsCollection
           .add({
@@ -154,90 +208,93 @@ export default {
             price: this.modifiedItem.price
           })
           .then(doc => {
-            this.modifiedItem.id = "";
-            this.modifiedItem.code = "";
-            this.modifiedItem.name = "";
-            this.modifiedItem.description = "";
-            this.modifiedItem.category = "";
-            this.modifiedItem.u_id = "";
-            this.modifiedItem.imageurl = "";
-            this.modifiedItem.price = "";
-            this.showAddsection = false;
-            this.isEdit = false;
+            this.modifiedItem.id = ''
+            this.modifiedItem.code = ''
+            this.modifiedItem.name = ''
+            this.modifiedItem.description = ''
+            this.modifiedItem.category = ''
+            this.modifiedItem.u_id = ''
+            this.modifiedItem.imageurl = ''
+            this.modifiedItem.price = ''
+            this.showAddsection = false
+            this.isEdit = false
           })
           .catch(err => {
-            console.log(err);
-          });
+            console.log(err)
+          })
       }
     },
 
-    editItem(item) {
-      this.isEdit = true;
-      this.modifiedItem = item;
-      this.showAddsection = true;
+    editItem (item) {
+      this.isEdit = true
+      this.modifiedItem = item
+      this.showAddsection = true
     },
-    deleteItem(id) {
+    deleteItem (id) {
       fb.itemsCollection
         .doc(id)
         .delete()
         .catch(err => {
-          console.log(err);
-        });
+          console.log(err)
+        })
     },
-    viewAddSection() {
-      this.showAddsection = true;
+    viewAddSection () {
+      this.showAddsection = true
     },
-    closePostModal() {
-      this.showAddsection = false;
+    closePostModal () {
+      this.showAddsection = false
     },
-    onPickFile() {
-      this.$refs.fileInput.click();
+    onPickFile () {
+      this.$refs.fileInput.click()
     },
-    onFilePicked(event) {
-      const fileReader = new FileReader();
-      fileReader.addEventListener("load", () => {
-        this.imageUrl = fileReader.result;
-        this.imageFile = event.target.files[0];
-      });
-      this.image = fileReader.readAsDataURL(event.target.files[0]);
+    onFilePicked (event) {
+      const fileReader = new FileReader()
+      fileReader.addEventListener('load', () => {
+        this.imageUrl = fileReader.result
+        this.imageFile = event.target.files[0]
+      })
+      this.image = fileReader.readAsDataURL(event.target.files[0])
     },
-    onImportFilePicked(event) {
-      var reader = new FileReader();
-      reader.readAsText(event.target.files[0]);
-      var fileinput;
+    onPickUploadFile () {
+      this.$refs.fileUpload.click()
+    },
+    onImportFilePicked (event) {
+      var reader = new FileReader()
+      reader.readAsText(event.target.files[0])
+      var fileinput
       reader.onload = e => {
-        fileinput = e.target.result;
-        var lines = fileinput.split("\n");
-        //Need to create a map as, by the time response comes from firebase, the data object gets modified by next iteration
-        var keyDataMap = new Object();
-        //Line 0 is header line
+        fileinput = e.target.result
+        var lines = fileinput.split('\n')
+        // Need to create a map as, by the time response comes from firebase, the data object gets modified by next iteration
+        var keyDataMap = {}
+        // Line 0 is header line
         for (var lineNumber = 1; lineNumber < lines.length; lineNumber++) {
-          var data = lines[lineNumber].split(",");
-          console.error("lineNumber" + lines[lineNumber].split(","));
-          if (data[0] != "") {
-            keyDataMap[data[0]] = data;
+          var data = lines[lineNumber].split(',')
+          console.error('lineNumber' + lines[lineNumber].split(','))
+          if (data[0] !== '') {
+            keyDataMap[data[0]] = data
             fb.itemsCollection
-              .where("code", "==", data[0])
+              .where('code', '==', data[0])
               .get()
               .then(querySnapshot => {
                 if (querySnapshot.docs.length > 0) {
                   console.error(
-                    "found" + keyDataMap[querySnapshot.docs[0].data().code]
-                  );
-                  var existingDoc = querySnapshot.docs[0].data();
+                    'found' + keyDataMap[querySnapshot.docs[0].data().code]
+                  )
+                  var existingDoc = querySnapshot.docs[0].data()
                   existingDoc.name =
-                    keyDataMap[querySnapshot.docs[0].data().code][1];
+                    keyDataMap[querySnapshot.docs[0].data().code][1]
                   existingDoc.description =
-                    keyDataMap[querySnapshot.docs[0].data().code][2];
+                    keyDataMap[querySnapshot.docs[0].data().code][2]
                   existingDoc.category =
-                    keyDataMap[querySnapshot.docs[0].data().code][3];
+                    keyDataMap[querySnapshot.docs[0].data().code][3]
                   existingDoc.price =
-                    keyDataMap[querySnapshot.docs[0].data().code][4];
+                    keyDataMap[querySnapshot.docs[0].data().code][4]
                   fb.itemsCollection
                     .doc(querySnapshot.docs[0].id)
-                    .set(existingDoc);
+                    .set(existingDoc)
                 } else {
-                  console.error("not found" + data);
+                  console.error('not found' + data)
                   fb.itemsCollection
                     .add({
                       u_id: uuid.v1(),
@@ -248,12 +305,12 @@ export default {
                       price: data[4]
                     })
                     .catch(err => {
-                      console.log(err);
-                    });
+                      console.log(err)
+                    })
                 }
-              });
+              })
           } else {
-            console.error("blank" + data);
+            console.error('blank' + data)
             fb.itemsCollection
               .add({
                 u_id: uuid.v1(),
@@ -264,12 +321,12 @@ export default {
                 price: data[5]
               })
               .catch(err => {
-                console.log(err);
-              });
+                console.log(err)
+              })
           }
         }
-      };
+      }
     }
   }
-};
+}
 </script>
