@@ -1,15 +1,17 @@
 <template>
   <div id="app">
-    <nav>
-        <router-link to='/users'>Manage users</router-link>
-        <router-link to='/categories'>Manage categories</router-link>
-        <router-link to='/items'>Manage items</router-link>
-    </nav>
-    <router-view />
+    <div>
+      <b-jumbotron bg-variant="info" text-variant="white" border-variant="dark">
+        <template slot="header">eCatalogue Management</template>
+        <template slot="lead">This is eCatalogue Management Application.</template>
+        <hr class="my-4">
+        <p>User can add and edit the master data using the application.</p>
+      </b-jumbotron>
+    </div>
+    <router-view/>
   </div>
 </template>
 
 <script>
-export default {
-}
+export default {}
 </script>
