@@ -1,34 +1,47 @@
 <template>
-    <div id="login">
+  <div id="login">
+    <transition name="fade">
+      <div v-if="performingRequest" class="loading">
+        <p>Loading...</p>
+      </div>
+    </transition>
+    <section>
+      <div class="col1">
+        <h1>eCatalogue Admin</h1>
+      </div>
+      <div class="col2">
+        <b-form @submit.prevent>
+          <h1>Welcome Back</h1>
+          <b-form-group label="Email:" label-for="email1">
+            <b-form-input
+              id="email1"
+              v-model="loginForm.email"
+              type="text"
+              required
+              placeholder="you@email.com"
+            ></b-form-input>
+          </b-form-group>
+
+          <b-form-group label="Password:" label-for="password1">
+            <b-form-input
+              id="password1"
+              v-model="loginForm.password"
+              type="password"
+              required
+              placeholder="******"
+            ></b-form-input>
+          </b-form-group>
+
+          <b-button center variant="outline-primary" @click="login" class="button">Log In</b-button>
+        </b-form>
         <transition name="fade">
-            <div v-if="performingRequest" class="loading">
-                <p>Loading...</p>
-            </div>
+          <div v-if="errorMsg !== ''" class="error-msg">
+            <p>{{ errorMsg }}</p>
+          </div>
         </transition>
-        <section>
-            <div class="col1">
-                <h1>eCatalogue Admin</h1>
-            </div>
-            <div class="col2">
-                <form  @submit.prevent>
-                    <h1>Welcome Back</h1>
-
-                    <label for="email1">Email</label>
-                    <input v-model.trim="loginForm.email" type="text" placeholder="you@email.com" id="email1" />
-
-                    <label for="password1">Password</label>
-                    <input v-model.trim="loginForm.password" type="password" placeholder="******" id="password1" />
-
-                    <button @click="login" class="button">Log In</button>
-                </form>
-                <transition name="fade">
-                    <div v-if="errorMsg !== ''" class="error-msg">
-                        <p>{{ errorMsg }}</p>
-                    </div>
-                </transition>
-            </div>
-        </section>
-    </div>
+      </div>
+    </section>
+  </div>
 </template>
 
 <script>
@@ -36,6 +49,7 @@ const fb = require('../firebaseConfig.js')
 export default {
   data () {
     return {
+      performingRequest: false,
       loginForm: {
         email: '',
         password: ''
@@ -51,15 +65,21 @@ export default {
     },
     login () {
       this.performingRequest = true
-      fb.auth.signInWithEmailAndPassword(this.loginForm.email, this.loginForm.password).then(user => {
-        this.$store.commit('setCurrentUser', user)
-        this.performingRequest = false
-        this.$router.push('/dashboard')
-      }).catch(err => {
-        console.log(err)
-        this.performingRequest = false
-        this.errorMsg = err.message
-      })
+      fb.auth
+        .signInWithEmailAndPassword(
+          this.loginForm.email,
+          this.loginForm.password
+        )
+        .then(user => {
+          this.$store.commit('setCurrentUser', user)
+          this.performingRequest = false
+          this.$router.push('/home')
+        })
+        .catch(err => {
+          console.log(err)
+          this.performingRequest = false
+          this.errorMsg = err.message
+        })
     }
   }
 }
